video: fsl-dcu-fb: support signal polarity configuration
Use the display timing configurations hsync-/vsync- and pixelclk-
active properties. The hsync- and vsync-active configuration is
available through the fb_videomode sync field. But pixelclk-active
need to be stored seperately because when converting the struct
display_timing to struct fb_videomode, this information gets lost.

Hence we cannot pick a mode from the modelist and expect that
pixelclk-active is set correctly. But currently modelist is not
used by anything, hence this is no issue.